ADD : WBS Task CRUD + Excel download MODIFY : Disunite Division in Radar chart

+++ src/main/java/com/example/smstest/domain/task/TaskCategoryRepository.java
@@ -2,6 +2,6 @@ package com.example.smstest.domain.task;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TaskCategoryRepository extends JpaRepository<TaskCategory, Long> {
+public interface TaskCategoryRepository extends JpaRepository<TaskCategory, Integer> {
 
 }
